backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
sqlalchemy==2.0.31
asyncpg==0.29.0
alembic==1.13.2
psycopg2-binary==2.9.9
openpyxl==3.1.5
pandas==2.2.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.8.2
pydantic-settings==2.3.4
aiofiles==23.2.1
httpx==0.27.0
